Pull "ephemeral" collections into their own module.
The meaning and usage of the various `SimpleWhatever` collections was
unclear with them colocated with the abstract implementation. Moving
them to their own `ephemeral` module and not having them in the main
`somacore` namespace will make them less attractive. They still have
value in being the most basic possible implementation of a collection,
and are useful in tests for this package.

In the process of removing the `SimpleCollection` export from the
package root, also changes the `__init__.py` to just import module
members rather than the module itself. Since this is at the top level
and is exclusively for re-export, this does not have the same potential
circular dependency problems as doing so within internal modules, while
reducing repetition and the possibility of errors.
