+++ python-spec/src/somacore/ephemeral/__init__.py
@@ -1,0 +1,16 @@
+"""In-memory-only implementations of SOMA types.
+
+These are meant for testing and exploration, where a user may want to do an
+ad-hoc analysis of multiple datasets without having to create a stored
+Collection.
+"""
+
+from .collections import Collection
+from .collections import Experiment
+from .collections import Measurement
+
+__all__ = (
+    "Collection",
+    "Experiment",
+    "Measurement",
+)
